BaytesCore

Documentation for BaytesCore.

BaytesCore.AbstractAlgorithm

BaytesCore.AbstractConfiguration

BaytesCore.AbstractConstructor

BaytesCore.AbstractDiagnostics

BaytesCore.AbstractKernel

BaytesCore.AbstractTune

BaytesCore.AcceptStatistic

BaytesCore.Accumulator

BaytesCore.ArrayConfig

BaytesCore.BaseDiagnostics

BaytesCore.Batch

BaytesCore.ByCols

BaytesCore.ByRows

BaytesCore.ChainsTune

BaytesCore.DataSorting

BaytesCore.DataStructure

BaytesCore.DataTune

BaytesCore.Expanding

BaytesCore.IterationTempering

BaytesCore.Iterator

BaytesCore.JitterTune

BaytesCore.JointTempering

BaytesCore.ModelParameterBuffer

BaytesCore.ParameterBuffer

BaytesCore.ParameterWeighting

BaytesCore.ParameterWeights

BaytesCore.PrintDefault

BaytesCore.ProgressReport

BaytesCore.ResampleTune

BaytesCore.ResamplingMethod

BaytesCore.Rolling

BaytesCore.SampleDefault

BaytesCore.SubSampled

BaytesCore.TemperingMethod

BaytesCore.TemperingParameter

BaytesCore.UpdateBool

BaytesCore.Updater

BaytesCore.ValueHolder

BaytesCore.Tuple_to_Namedtuple

BaytesCore.adjust

BaytesCore.bisect

BaytesCore.checktemperature

BaytesCore.computeESS

BaytesCore.draw!

BaytesCore.generate_showvalues

BaytesCore.get_ESSDifference

BaytesCore.get_prediction

BaytesCore.get_result

BaytesCore.get_sym

BaytesCore.get_tagged

BaytesCore.grab

BaytesCore.infer

BaytesCore.init

BaytesCore.init

BaytesCore.init!

BaytesCore.initial

BaytesCore.initial

BaytesCore.issmaller

BaytesCore.jitter!

BaytesCore.keyunion

BaytesCore.logmeanexp

BaytesCore.normalize!

BaytesCore.propagate

BaytesCore.propagate!

BaytesCore.propose

BaytesCore.propose!

BaytesCore.randcat

BaytesCore.resample!

BaytesCore.result!

BaytesCore.results

BaytesCore.shuffle_backward!

BaytesCore.shuffle_forward!

BaytesCore.subset

BaytesCore.subset

BaytesCore.to_NamedTuple

BaytesCore.to_NamedTuple_generated

BaytesCore.to_Tuple

BaytesCore.to_Tuple_generated

BaytesCore.to_Tuple_generated

BaytesCore.untemper

BaytesCore.update

BaytesCore.update

BaytesCore.update

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.weight!

BaytesCore.weightedESS

BaytesCore.weightedincrement

Random.shuffle!

BaytesCore.AbstractAlgorithm — Type

abstract type AbstractAlgorithm

Abstract super type for Algorithms. New algorithm needs to be subtype of AbstractAlgorithm.

BaytesCore.AbstractConfiguration — Type

abstract type AbstractConfiguration

Lower level configuration for individual Algorithms.

BaytesCore.AbstractConstructor — Type

abstract type AbstractConstructor

Abstract type to construct Algorithms. New algorithm needs to define a functor to initiate corresponding AbstractAlgorithm.

BaytesCore.AbstractDiagnostics — Type

abstract type AbstractDiagnostics

Abstract super type for Algorithm Diagnostics. Once AbstractAlgorithm has run, return new model parameter and AbstractDiagnostics of AbstractAlgorithm.

BaytesCore.AbstractKernel — Type

abstract type AbstractKernel

Abstract super type for Algorithm Kernels. New Kernels needs to be inside corresponding AbstractAlgorithm.

BaytesCore.AbstractTune — Type

abstract type AbstractTune

Abstract super type for Tuning Algorithms.

BaytesCore.AcceptStatistic — Type

struct AcceptStatistic

Stores in diagnostics.

Fields

rate::Float64

Acceptance rate

accepted::Bool

Step accepted or rejected

BaytesCore.Accumulator — Type

mutable struct Accumulator

Mutable iterator struct that can be stored and updated in immutable structs. Contains cumulative and current value of input.

Fields

cumulative::Float64

Cumulative value of all inputs.

current::Float64

Current input value.

BaytesCore.ArrayConfig — Type

struct ArrayConfig{K, T<:BaytesCore.DataSorting}

Array configuration struct.

Checks if data is propagated row or column wise.

Fields

size::Tuple{Vararg{Int64, K}} where K

Dimension of data.

sorted::BaytesCore.DataSorting

Boolean if data is sorted by row

BaytesCore.BaseDiagnostics — Type

struct BaseDiagnostics{P}

Stores summary information for diagnostics output that is included in all Baytes kernels.

Fields

ℓobjective::Float64

Log objective target result.

temperature::Float64

Temperature for target result

prediction::Any

Prediction of future data.

iter::Int64

Current iteration of kernel.

BaytesCore.Batch — Type

struct Batch <: DataStructure

Determines whether only full data can be used.

Fields

BaytesCore.ByCols — Type

Data is sorted by columns.

BaytesCore.ByRows — Type

Data is sorted by rows.

BaytesCore.ChainsTune — Type

mutable struct ChainsTune

Holds tuning information for number of chains in comparison to data.

Fields

coverage::Float64

Proposed coverage of number of chains/number of data points.

threshold::Float64

Threshold for resampling, between 0.0 and 1.0.

Nchains::Int64

Number of chains.

Ndata::Int64

Number of data points.

BaytesCore.DataSorting — Type

Abstract super type for sorting data in different dimensions.

BaytesCore.DataStructure — Type

Abstract super type to define valid data structures.

BaytesCore.DataTune — Type

struct DataTune{D<:DataStructure, C<:Union{Nothing, ArrayConfig}, M}

Data tuning struct.

Contains information about data structure.

Fields

structure::DataStructure

Subtype of DataStructure

config::Union{Nothing, ArrayConfig}

Data dimension configuration.

miss::Any

Treatment of missing data.

BaytesCore.Expanding — Type

struct Expanding <: DataStructure

Determines if data size for sampler is increasing over time.

Fields

index::Iterator

maximal visible index

BaytesCore.IterationTempering — Type

struct IterationTempering{A<:UpdateBool, T<:AbstractFloat} <: TemperingMethod

Tuning container for iteration uptdates of temperature.

Fields

adaption::UpdateBool

Checks if temperature will be updated after new iterations

initial::ValueHolder{T} where T<:AbstractFloat

Initial temperature for tempering.

parameter::TemperingParameter

Tuning parameter for temperature adjustment.

BaytesCore.Iterator — Type

mutable struct Iterator

Mutable struct that can be stored and updated in immutable structs. Contains current iteration value.

Fields

current::Int64

Current iteration number.

BaytesCore.JitterTune — Type

struct JitterTune{B<:UpdateBool}

Container that determines if jittering has to be applied

Fields

adaption::UpdateBool

Bolean if adaptive jittering is applied. If false, always max steps are jittered.

Nsteps::Iterator

Number of jitttering steps performed while jittering

threshold::Float64

Jittering threshold for correlation of parameter.

min::Int64

Minimum amount of jittering steps.

max::Int64

Maximum amount of jittering steps.

BaytesCore.JointTempering — Type

struct JointTempering{A<:UpdateBool, T<:AbstractFloat} <: TemperingMethod

Tuning container for joint uptdates of temperature.

Fields

adaption::UpdateBool

Checks if temperature will be updated after new iterations

val::ValueHolder{T} where T<:AbstractFloat

current temperature.

ESSTarget::ValueHolder{T} where T<:AbstractFloat

Target Effective Sample Size.

weights::Vector{T} where T<:AbstractFloat

Buffer for weights used to adapt temperature

BaytesCore.ModelParameterBuffer — Type

Stores a vector of a all model parameter val and several useful location statistics in the chain.

Examples

BaytesCore.ParameterBuffer — Type

Stores a vector of a single model parameter val and its current index index in the chain.

Examples

BaytesCore.ParameterWeighting — Type

abstract type ParameterWeighting

Super type for parameter weighting techniques.

BaytesCore.ParameterWeights — Type

struct ParameterWeights

Container for weights and normalized particles weights at current time step t.

Fields

ℓweights::Vector{Float64}

Used for log likelihood calculation.

ℓweightsₙ::Vector{Float64}

exp(ℓweightsₙ) used for resampling particles and sampling trajectories ~ kept in log space for numerical stability.

buffer::Vector{Float64}

A buffer vector with same size as ℓweights and ℓweightsₙ.

BaytesCore.PrintDefault — Type

struct PrintDefault

Default arguments for printing summary statistics after sampling.

Fields

Ndigits::Int64

Number of Digits for rounding summary statistics.

quantiles::Vector{Float64}

Quantiles for summary statistics.

BaytesCore.ProgressReport — Type

struct ProgressReport{P<:ProgressLog, K}

Default arguments for logging sampling output during sampling process.

Fields

bar::Bool

Boolean if progressbar enabled during sampling run

log::ProgressLog

Boolean if output diagnostics should be printed.

kwargs::Any

Additional ProgressMeter keywords to be used to return output

BaytesCore.ResampleTune — Type

struct ResampleTune{R<:ResamplingMethod}

Stores information about resampling steps.

Fields

method::ResamplingMethod

Method for resampling.

update::Updater

Stores if last iteration was resampled.

BaytesCore.ResamplingMethod — Type

abstract type ResamplingMethod

Super type for various particle resampling techniques.

BaytesCore.Rolling — Type

struct Rolling <: DataStructure

Determines if data size for sampler is rolled forward over time.

Fields

length::Int64

Rolling window.

index::Iterator

maximal visible index

BaytesCore.SampleDefault — Type

struct SampleDefault{D<:DataStructure, M<:TemperingMethod, P<:ProgressReport}

Default arguments for sampling routine.

Fields

dataformat::DataStructure

Format to split data before proposal step.

tempering::TemperingMethod

Tempering values for target distributions of each chain, per default = 1*logtarget(θ | data) for each chain

chains::Int64

Number of chains run in sampling step.

iterations::Int64

Number of MCMC iterations. May be overwritten in SMC.

burnin::Int64

Burn in samples.

thinning::Int64

Number of consecutive samples taken for diagnostics output.

safeoutput::Bool

Boolean if trace and algorithm should be safed to working directory.

printoutput::Bool

Boolean if summary statistics should be calculated.

printdefault::PrintDefault

Default arguments for printing summary statistics

report::ProgressReport

Default arguments for printing output during sampling runs

BaytesCore.SubSampled — Type

struct SubSampled <: DataStructure

Determines whether data can be subsampled for inference.

Fields

buffer::Vector{Int64}

Buffer index for subsampled indices.

BaytesCore.TemperingMethod — Type

Abstract type where more tempering methods have to be dispatched on.

BaytesCore.TemperingParameter — Type

struct TemperingParameter{T<:AbstractFloat}

Holds information about updating tempering parameter, parameter for basic logistic function.

Fields

L::AbstractFloat

k::AbstractFloat

x₀::AbstractFloat

BaytesCore.UpdateBool — Type

abstract type UpdateBool

Abstract super type if update has to be applied for given function.

Fields

BaytesCore.Updater — Type

mutable struct Updater

Mutable struct to store update information that can be stored and updated in immutable structs. Contains boolean if updated is needed. Note that a UpdateBool field cannot be stored and updated as a concrete type here.

Fields

current::Bool

Boolean of whether update has to be applied

BaytesCore.ValueHolder — Type

mutable struct ValueHolder{T<:Real}

Mutable struct to hold a single scalar value that can be stored and updated in immutable structs.

Fields

current::Real

Scalar value

BaytesCore.Tuple_to_Namedtuple — Method

Convert Tuple tup to NamedTuple with all fields equal to val.

BaytesCore.adjust — Method

adjust(tune, data)

Adjust data with current datatune configuration.

Examples

BaytesCore.bisect — Function

Inverse CDF of Univariate Normal Mixture Model, where prob can be vectorized as a bivariate array of observations.

BaytesCore.checktemperature — Method

checktemperature(param, iterations)

Helper function to determine temperature shape given param variables.

Examples

BaytesCore.computeESS — Method

computeESS(weightsₙ)

Stable version of computing effectice sample size of particle filter via normalized log weights.

Examples

BaytesCore.draw! — Method

draw!(_rng, weights)

Draw proposal trajectory index.

Examples

BaytesCore.generate_showvalues — Function

generate_showvalues

Show values of Algorihm diagnostics.

Examples

generate_showvalues(diagnostics::AbstractDiagnostics)

BaytesCore.get_ESSDifference — Method

get_ESSDifference(weights, λₜ₋₁, ESSTarget)

Closure to compute new temperature λ as a function of weights weights and current temperature λₜ₋₁. ESSTarget is the benchmark to optimize against.

Examples

BaytesCore.get_prediction — Method

get_prediction

Return prediction of AbstractAlgorithm diagnostics.

BaytesCore.get_result — Function

get_result

Show log density result of algorithm.

Examples

get_result(algorithm::AbstractAlgorithm)

BaytesCore.get_sym — Method

get_sym

Return tagged symbol of AbstractConstructor.

BaytesCore.get_tagged — Method

get_tagged

Get tagged parameter of AbstractAlgorithm..

BaytesCore.grab — Method

grab(data, index, sorted)

Return allocation friendly index of array argument.

Examples

BaytesCore.infer — Function

infer

Infer output type from input. Infer AbstractDiagnostics from corresponding AbstractAlgorithm

Examples

infer(_rng, diagnostics::Type{AbstractDiagnostics}, algorithm, model, data)

BaytesCore.init — Function

init

Initialize struct.

BaytesCore.init! — Function

init!

Inplace version of init.

BaytesCore.init — Method

init(_, adaption, val, parameter)

Initialize temperature.

Examples

BaytesCore.initial — Method

initial(tempertune)

Return initial temperature of tuning container.

Examples

BaytesCore.initial — Method

initial(tempertune)

Return initial temperature of tuning container.

Examples

BaytesCore.issmaller — Method

issmaller(x, threshold)

Check if x is larger than threshold, guaranteeing same type for arguments.

Examples

BaytesCore.jitter! — Method

jitter!(tune, adaption, ρ)

Check if we can stop jittering step based on correlation ρ.

Examples

BaytesCore.keyunion — Method

Generate union of fields from 2 NamedTuples x and y.

BaytesCore.logmeanexp — Method

logmeanexp(arr)

Numerically stable version for log(mean(exp(vec))).

Examples

BaytesCore.normalize! — Method

normalize!(weights)

Inplace-Normalize parameter weights, accounting for ℓweightsₙ at previous iteration.

Examples

BaytesCore.propagate — Function

propagate

Propagate algorithm one step forward.

BaytesCore.propagate! — Function

propagate!

Inplace version of propagate.

BaytesCore.propose — Function

propose

Propose new parameter with given algorithm, keeping objective fixed.

Examples

propose(_rng::AbstractRNG, algorithm::AbstractAlgorithm, objective::AbstractObjective)

BaytesCore.propose! — Function

propose!

Inplace version of propose.

Examples

propose!(_rng::AbstractRNG, algorithm::AbstractAlgorithm, model::AbstractModelWrapper, data, temperature, update)

BaytesCore.randcat — Method

randcat(_rng, p)

More stable, faster version of rand(Categorical) if weights sum up to 1.

Examples

BaytesCore.resample! — Function

Resample particles, dispatched on ResamplingMethod subtypes.

Examples

BaytesCore.result! — Function

result!

Substitute result position in algorithm with new result.

Examples

result!(algorithm::AbstractAlgorithm, result::AbstractResult)

BaytesCore.results — Function

results

Show results of input type, in this case a vector of AbstractDiagnostics of corresponding AbstractAlgorithm.

Examples

results(diagnosticsᵛ::AbstractVector{AbstractDiagnostics}, algorithm, Ndigits, quantiles)

BaytesCore.shuffle_backward! — Method

shuffle_backward!(val, buffer_val, ancestor, buffer_ancestor)

Similar to suffle, sort val elements at last column according to buffer indices, using ancestors as buffer. buffer indices will be different for each column, and will be adjusted accordingly.

Examples

BaytesCore.shuffle_forward! — Method

shuffle_forward!(val, buffer_val, ancestor, buffer_ancestor, lookback, iterₘₐₓ)

Sort val elements from iterₘₐₓ up until 'lookback' according to ancestor indices at iterₘₐₓ. ancestor at `iterₘₐₓ - lookback' will be adjusted accordingly.

Examples

BaytesCore.subset — Method

Subset NamedTuple obj with keys of s without allocations.

BaytesCore.subset — Method

subset(nt, s)

Subset nt with arguments s

Examples

julia> subset( (a = 1., b = 2., c = 3.), (:a, :b) )
(a = 1., b = 2.)

BaytesCore.to_NamedTuple — Method

Convert Tuple tup to NamedTuple with fields equal to val.

BaytesCore.to_NamedTuple_generated — Method

Convert struct x to NamedTuple without additional allocations.

BaytesCore.to_Tuple — Method

to_Tuple(sym)

Return symbol input as tuple.

Examples

BaytesCore.to_Tuple_generated — Method

Subset NamedTuple x given symbols sym without additional allocations.

BaytesCore.to_Tuple_generated — Method

Convert strcut x to tuple.

BaytesCore.untemper — Method

untemper(diagnostics)

Return ℓobjective scaled up to temperature == 1.

Examples

BaytesCore.update — Function

update

Update struct and return new type.

BaytesCore.update! — Function

update!

Update fields of existing struct.

BaytesCore.update! — Method

update!(weights)

Set weights back to equal.

Examples

BaytesCore.update! — Method

update!(_rng, tune, structure)

Update datetune struct based on datastructure.

Examples

BaytesCore.update — Method

Temperature adaption that takes into account updating schedule, see Chopin, Papaspiliopoulos (2020).

Examples

BaytesCore.update — Method

update(param, index)

Calculate new temperature based on current iteration.

Examples

BaytesCore.weight! — Method

weight!()

Weight function that is dispatched on BaytesCore.ParameterWeighting types.

Examples

BaytesCore.weightedESS — Method

Compute ESS as a function of proposed temperature λ against current temperature λₜ₋₁. Note that weights are in the original (non-log) space, and should have temperature == 1.

Examples

BaytesCore.weightedincrement — Method

weightedincrement(weights)

Stable version of computing weighted incremental weight average.

Examples

Random.shuffle! — Method

shuffle!(val, buffer, ancestor)

Sort val elements at last column according to buffer indices, using ancestors as buffer. Resort all previous columns accordingly.

Examples