BaytesCore

Documentation for BaytesCore.

BaytesCore.AbstractAlgorithm

BaytesCore.AbstractConfiguration

BaytesCore.AbstractConstructor

BaytesCore.AbstractDiagnostics

BaytesCore.AbstractKernel

BaytesCore.AbstractTune

BaytesCore.AcceptStatistic

BaytesCore.Accumulator

BaytesCore.ArrayConfig

BaytesCore.Batch

BaytesCore.ByCols

BaytesCore.ByRows

BaytesCore.DataSorting

BaytesCore.DataStructure

BaytesCore.DataTune

BaytesCore.Expanding

BaytesCore.Iterator

BaytesCore.JitterTune

BaytesCore.ParameterWeighting

BaytesCore.ParameterWeights

BaytesCore.PrintDefault

BaytesCore.ResamplingMethod

BaytesCore.Rolling

BaytesCore.SubSampled

BaytesCore.TemperDefault

BaytesCore.TemperingParameter

BaytesCore.TemperingTune

BaytesCore.UpdateBool

BaytesCore.Updater

BaytesCore.ValueHolder

Base.split

BaytesCore.Tuple_to_Namedtuple

BaytesCore.adapt

BaytesCore.adapt!

BaytesCore.adjust

BaytesCore.checktemperature

BaytesCore.computeESS

BaytesCore.draw!

BaytesCore.generate_showvalues

BaytesCore.get_iteration

BaytesCore.get_loglik

BaytesCore.get_phase

BaytesCore.get_prediction

BaytesCore.get_result

BaytesCore.get_tagged

BaytesCore.grab

BaytesCore.infer

BaytesCore.init

BaytesCore.init

BaytesCore.init!

BaytesCore.islarger

BaytesCore.jitter!

BaytesCore.keyunion

BaytesCore.logmeanexp

BaytesCore.normalize!

BaytesCore.propagate

BaytesCore.propagate!

BaytesCore.propose

BaytesCore.propose!

BaytesCore.randcat

BaytesCore.resample!

BaytesCore.result!

BaytesCore.results

BaytesCore.shuffle_backward!

BaytesCore.shuffle_forward!

BaytesCore.subset

BaytesCore.subset

BaytesCore.to_NamedTuple

BaytesCore.to_NamedTuple_generated

BaytesCore.to_Tuple_generated

BaytesCore.to_Tuple_generated

BaytesCore.update

BaytesCore.update

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.weight!

Random.shuffle!

BaytesCore.AbstractAlgorithm — Type

abstract type AbstractAlgorithm

Abstract super type for Algorithms.

BaytesCore.AbstractConfiguration — Type

abstract type AbstractConfiguration

Lower level configuration for individual Algorithms.

BaytesCore.AbstractConstructor — Type

abstract type AbstractConstructor

Abstract type to construct Algorithms.

BaytesCore.AbstractDiagnostics — Type

abstract type AbstractDiagnostics

Abstract super type for Algorithm Diagnostics.

BaytesCore.AbstractKernel — Type

abstract type AbstractKernel

Abstract super type for Algorithm Kernels.

BaytesCore.AbstractTune — Type

abstract type AbstractTune

Abstract super type for Tuning Algorithms.

BaytesCore.AcceptStatistic — Type

struct AcceptStatistic{T<:AbstractFloat}

Stores in diagnostics.

Fields

rate::AbstractFloat

Acceptance rate

accepted::Bool

Step accepted or rejected

BaytesCore.Accumulator — Type

mutable struct Accumulator

Mutable iterator struct that can be stored and updated in immutable structs. Contains cumulative and current value of input.

Fields

cumulative::Float64

Cumulative value of all inputs.

current::Float64

Current input value.

BaytesCore.ArrayConfig — Type

struct ArrayConfig{K, T<:BaytesCore.DataSorting}

Array configuration struct.

Checks if data is propagated row or column wise.

Fields

size::Tuple{Vararg{Int64, K}} where K

Dimension of data.

sorted::BaytesCore.DataSorting

Boolean if data is sorted by row

BaytesCore.Batch — Type

struct Batch <: DataStructure

Determines whether only full data can be used.

Fields

BaytesCore.ByCols — Type

Data is sorted by columns.

BaytesCore.ByRows — Type

Data is sorted by rows.

BaytesCore.DataSorting — Type

Abstract super type for sorting data in different dimensions.

BaytesCore.DataStructure — Type

Abstract super type to define valid data structures.

BaytesCore.DataTune — Type

struct DataTune{D<:DataStructure, C<:ArrayConfig, M}

Data tuning struct.

Contains information about data structure.

Fields

structure::DataStructure

Subtype of DataStructure

config::ArrayConfig

Data dimension configuration.

miss::Any

Treatment of missing data.

BaytesCore.Expanding — Type

struct Expanding <: DataStructure

Determines if data size for sampler is increasing over time.

Fields

index::Iterator

maximal visible index

BaytesCore.Iterator — Type

mutable struct Iterator

Mutable struct that can be stored and updated in immutable structs. Contains current iteration value.

Fields

current::Int64

Current iteration number.

BaytesCore.JitterTune — Type

struct JitterTune

Container that determines if jittering has to be applied

Fields

Nsteps::Iterator

Number of jitttering steps performed while jittering

threshold::Float64

Jittering threshold for correlation of parameter.

min::Int64

Minimum amount of jittering steps.

max::Int64

Maximum amount of jittering steps.

BaytesCore.ParameterWeighting — Type

abstract type ParameterWeighting

Super type for parameter weighting techniques.

BaytesCore.ParameterWeights — Type

struct ParameterWeights

Container for weights and normalized particles weights at current time step t.

Fields

ℓweights::Vector{Float64}

Used for log likelihood calculation.

ℓweightsₙ::Vector{Float64}

exp(ℓweightsₙ) used for resampling particles and sampling trajectories ~ kept in log space for numerical stability.

buffer::Vector{Float64}

A buffer vector with same size as ℓweights and ℓweightsₙ.

BaytesCore.PrintDefault — Type

struct PrintDefault

Default arguments for printing summary statistics after sampling.

Fields

Ndigits::Int64

Number of Digits for rounding summary statistics.

quantiles::Vector{Float64}

Quantiles for summary statistics.

BaytesCore.ResamplingMethod — Type

abstract type ResamplingMethod

Super type for various particle resampling techniques.

BaytesCore.Rolling — Type

struct Rolling <: DataStructure

Determines if data size for sampler is rolled forward over time.

Fields

length::Int64

Rolling window.

index::Iterator

maximal visible index

BaytesCore.SubSampled — Type

struct SubSampled <: DataStructure

Determines whether data can be subsampled for inference.

Fields

buffer::Vector{Int64}

Buffer index for subsampled indices.

BaytesCore.TemperDefault — Type

struct TemperDefault{A<:UpdateBool, T<:(Union{AbstractVector{R}, R} where R<:AbstractFloat)}

Default arguments for tempering target function.

Fields

adaption::UpdateBool

Checks if temperature will be updated after new iterations

val::Union{AbstractVector{R}, R} where R<:AbstractFloat

Initial value for temperature if adaption is used. Otherwise used to change shape of temperature until it reaches 1.0. The smaller, the more linear the trajectory. The higher, the sooner 1.0 will be reached.

BaytesCore.TemperingParameter — Type

struct TemperingParameter{T<:AbstractFloat}

Holds information about updating tempering parameter, parameter for basic logistic function

Fields

L::AbstractFloat

k::AbstractFloat

x₀::AbstractFloat

BaytesCore.TemperingTune — Type

struct TemperingTune{A<:UpdateBool, T<:AbstractFloat}

Summary struct for Tempering parameter.

Fields

adaption::UpdateBool

If true, temperature will be adapted.

val::ValueHolder{T} where T<:AbstractFloat

current temperature.

parameter::TemperingParameter

Tuning parameter for temperature adjustment.

BaytesCore.UpdateBool — Type

abstract type UpdateBool

Abstract super type if update has to be applied for given function.

Fields

BaytesCore.Updater — Type

mutable struct Updater

Mutable struct to store update information that can be stored and updated in immutable structs. Contains boolean if updated is needed. Note that a UpdateBool field cannot be stored and updated as a concrete type here.

Fields

current::Bool

Boolean of whether update has to be applied

BaytesCore.ValueHolder — Type

mutable struct ValueHolder{T<:Real}

Mutable struct to hold a single scalar value that can be stored and updated in immutable structs.

Fields

current::Real

Scalar value

Base.split — Method

Split default.val temperatures into sepatre TemperDefault structs with scalar temperature.

BaytesCore.Tuple_to_Namedtuple — Method

Convert Tuple tup to NamedTuple with all fields equal to val.

BaytesCore.adapt — Function

adapt

Update tuning container

BaytesCore.adapt! — Function

adapt!

Inplace version of adapt.

BaytesCore.adjust — Method

adjust(tune, data)

Adjust data with current datatune configuration.

Examples

BaytesCore.checktemperature — Method

checktemperature(param, iterations)

Helper function to determine temperature shape given param variables.

Examples

BaytesCore.computeESS — Method

computeESS(weightsₙ)

Compute effectice sample size of particle filter via normalized log weights.

Examples

BaytesCore.draw! — Method

draw!(_rng, weights)

Draw proposal trajectory index.

Examples

BaytesCore.generate_showvalues — Function

generate_showvalues

Show values of Algorihm diagnostics.

BaytesCore.get_iteration — Function

get_iteration

Get current iteration of sampler combination.

BaytesCore.get_loglik — Function

get_loglik

Get log likelihood of sampler combination.

BaytesCore.get_phase — Function

get_phase

Get phase tune struct of sampler combination.

BaytesCore.get_prediction — Function

get_prediction

Get predictions of sampler combination.

BaytesCore.get_result — Function

get_result

Show log density result.

BaytesCore.get_tagged — Function

get_tagged

Show tagged parameter.

BaytesCore.grab — Method

grab(data, index, sorted)

Return allocation friendly index of array argument.

Examples

BaytesCore.infer — Function

infer

Infer output type from input

BaytesCore.init — Function

init

Initialize struct.

BaytesCore.init! — Function

init!

Inplace version of init.

BaytesCore.init — Method

init(_, adaption, default, parameter)

Initialize temperature.

Examples

BaytesCore.islarger — Method

islarger(x, threshold)

Check if x is larger than threshold, guaranteeing same type for arguments.

Examples

BaytesCore.jitter! — Method

jitter!(tune, ρ)

Check if we can stop jittering step based on correlation ρ.

Examples

BaytesCore.keyunion — Method

Generate union of fields from 2 NamedTuples x and y.

BaytesCore.logmeanexp — Method

logmeanexp(arr)

Numerically stable version for log(mean(exp(vec))).

Examples

BaytesCore.normalize! — Method

normalize!(weights)

Inplace-Normalize parameter weights, accounting for ℓweightsₙ at previous iteration.

Examples

BaytesCore.propagate — Function

propagate

Propagate algorithm one step forward.

BaytesCore.propagate! — Function

propagate!

Inplace version of propagate.

BaytesCore.propose — Function

propose

Propose new parameter with given algorithm.

BaytesCore.propose! — Function

propose!

Inplace version of propose.

BaytesCore.randcat — Method

randcat(_rng, p)

More stable, faster version of rand(Categorical) if weights sum up to 1.

Examples

BaytesCore.resample! — Function

Resample particles, dispatched on ResamplingMethod subtypes.

Examples

BaytesCore.result! — Function

result!

Substitute result with new result.

BaytesCore.results — Function

results

Show results of input type.

BaytesCore.shuffle_backward! — Method

shuffle_backward!(val, buffer_val, ancestor, buffer_ancestor)

Similar to suffle, sort val elements at last column according to buffer indices, using ancestors as buffer. buffer indices will be different for each column, and will be adjusted accordingly.

Examples

BaytesCore.shuffle_forward! — Method

shuffle_forward!(val, buffer_val, ancestor, buffer_ancestor, lookback, iterₘₐₓ)

Sort val elements from iterₘₐₓ up until 'lookback' according to ancestor indices at iterₘₐₓ. ancestor at `iterₘₐₓ - lookback' will be adjusted accordingly.

Examples

BaytesCore.subset — Method

Subset NamedTuple obj with keys of s without allocations.

BaytesCore.subset — Method

subset(nt, s)

Subset nt with arguments s

Examples

julia> subset( (a = 1., b = 2., c = 3.), (:a, :b) )
(a = 1., b = 2.)

BaytesCore.to_NamedTuple — Method

Convert Tuple tup to NamedTuple with fields equal to val.

BaytesCore.to_NamedTuple_generated — Method

Convert struct x to NamedTuple without additional allocations.

BaytesCore.to_Tuple_generated — Method

Subset NamedTuple x given symbols sym without additional allocations.

BaytesCore.to_Tuple_generated — Method

Convert strcut x to tuple.

BaytesCore.update — Function

update

Update struct and return new type.

BaytesCore.update! — Function

update!

Update fields of existing struct.

BaytesCore.update! — Method

update!(weights)

Set weights back to equal.

Examples

BaytesCore.update! — Method

update!(_rng, tune, structure)

Update datetune struct based on datastructure.

Examples

BaytesCore.update — Method

update(param, index)

Calculate new temperature based on current iteration.

Examples

BaytesCore.weight! — Method

weight!()

Weight function that is dispatched on BaytesCore.ParameterWeighting types.

Examples

Random.shuffle! — Method

shuffle!(val, buffer, ancestor)

Sort val elements at last column according to buffer indices, using ancestors as buffer. Resort all previous columns accordingly.

Examples